begin aligned memory
Beginning aligned memory operations. We will have to modify the drivers only insofar as we need accesss to int NB (# of rows of B), which I think might best be accomplished by adding another variable to the struct def of the rotation plans, rather than computing it within every driver, as has been done in the permute functions.