start vectorized recurrence summation
The template for Horner's rule is:

ft_horner(n, c, incc, m, x, f)

For Clenshaw's algorithm, it's

ft_clenshaw(n, c, incc, m, x, f)

The coefficients have stride incc, but the points and the output array must be contiguous.

The points and output pointers may be aliased so that in-place evaluation can work with one array for x and f.

The design is as follows:
1. Find the native SIMD flags on the machine that is compiling the code.
2. Define as many internal functions using the highest level of SIMD with SIMD flags. Otherwise, define fallbacks.
3. Define the exported functions with SIMD dispatch based on GCC's cpuid.

This approach allows one to cross-compile and generate optimal code for machines that are as good as the cross compiler. It also allows the code to be callable from machines that are newer than the cross compiler (in case the cross compiler is not state-of-the-art) due to the fallbacks, though newer SIMD is not accessible.