Spin-weighted spherical harmonics and refactored use of SIMD (#43)
* start vectorized recurrence summation

The template for Horner's rule is:

ft_horner(n, c, incc, m, x, f)

For Clenshaw's algorithm, it's

ft_clenshaw(n, c, incc, m, x, f)

The coefficients have stride incc, but the points and the output array must be contiguous.

The points and output pointers may be aliased so that in-place evaluation can work with one array for x and f.

The design is as follows:
1. Find the native SIMD flags on the machine that is compiling the code.
2. Define as many internal functions using the highest level of SIMD with SIMD flags. Otherwise, define fallbacks.
3. Define the exported functions with SIMD dispatch based on GCC's cpuid.

This approach allows one to cross-compile and generate optimal code for machines that are as good as the cross compiler. It also allows the code to be callable from machines that are newer than the cross compiler (in case the cross compiler is not state-of-the-art) due to the fallbacks, though newer SIMD is not accessible.

* attempt to fix appveyor

* attempt to fix linux builds

The reported error is:

/usr/bin/ld: /tmp/ccuNTFaV.o: relocation R_X86_64_PC32 against undefined symbol `memset@@GLIBC_2.2.5' can not be used when making a shared object; recompile with -fPIC
500/usr/bin/ld: final link failed: Bad value

* windows fix bis

* link too?

* add tests for AVX512F

apparently, linux build machines already have it

* check SIMD checks

* AVX512F is an extended feature, which needs to be checked with __get_cpuid_count(7, ...)

* attempt cpuid

* fix spacing

* temporarily remove the recurrence module

* separate assembly from source in an attempt to get library flags closer to source?

* flatten the directory tree?

* specify the assembly wihout *

* gotta pump up those levels

* Le commencement du grand dérangement

* double4 => double8

* restore all drivers tests

* make the addition theorem work for any degree-n Legendre polynomial

not just degree 4

* cleanup execute_sph and kernel_sph

* does this hack fix the bug?

* fftw_execute_r2r => fftw_execute_dft_c2r

The dft_c2r obviates the need for a global transpose (in colswap) outside the fftw call. Instead, it requires setting and extracting some data from fftw_complex arrays.

* complex.h is apparently not always loaded in cblas.h

* add spin-weighted spherical harmonic transforms!

* try to remove most warnings across gcc versions & platforms

up error tolerance for triangular transforms for Windows

* ok better fix

* revert fix

no warning in gcc-8 is a warning in gcc-7 and vice versa

* refactor disk harmonic drivers!

* remove commented code

* finish cleaning up test_drivers and test_fftw

* add spin-weighted example

Add complex conjugation to docs

remove deploy binaries

* the algorithms work with N x M arrays, therefore the docs should reflect that the degrees are all N-1

\dh unrecognized => tab complete

* hide SIMD tetrahedral prototypes from interface

* update perf benchmark

* add with