add symmetric_tridiagonal_symmetric_eigen
This creates a structured form of the symmetric eigenvectors of a symmetric tridiagonal matrix with known eigenvalues.

The eigenvectors must be symmetric (this is not implied by the symmetric tridiagonal matrix, rather is something the user must known a priori) to calculate the normalization in only O(n) (rather than O(n^2)).

The `const int sign` determines the starting sign (sign of the [0,0] entry) of the eigenvectors.